feat: 메뉴 상세설명 구현

+++ app/src/main/java/com/example/kiosk/data/entity/Entity.java
@@ -1,5 +1,5 @@
 package com.example.kiosk.data.entity;
 
 public abstract class Entity {
-    public abstract int id();
+    public abstract long id();
 }
